pyproject.toml
[project]
name = "docxlib"
version = "0.1.0"
description = "A simple and easy-to-use Word document processing library based on Spire.Doc"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "DocxLib Contributors"}
]
keywords = ["word", "docx", "document", "spire", "office"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "spire-doc-free>=12.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]
excel = [
    "pandas>=1.3.0",
    "openpyxl>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/docxlib"
Repository = "https://github.com/yourusername/docxlib"
Documentation = "https://github.com/yourusername/docxlib/blob/main/README.md"
"Bug Tracker" = "https://github.com/yourusername/docxlib/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["docxlib"]

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
